Refactor contact controller and add COLREGS tests
